pyproject.toml
[project]
name = "okx-onchainos-contest-bot"
version = "0.1.0"
description = "Base-only trading bot for OKX OnchainOS AI competition"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "web3>=7.0.0",
  "eth-account>=0.13.0",
]

[project.scripts]
okx-contest-bot = "okx_contest_bot.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
